05-11: wallet transaction impl

+++ wallet/src/main/java/com/kulsin/wallet/balance/BalanceResponse.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class BalanceResponse {
 
     private final Long playerId;
-    private final Long balance;
+    private final double balance;
 
 }
